requirements_hr39.txt
absl-py==2.0.0
accelerate==0.25.0
aicsimageio==4.14.0
aicspylibczi==3.2.1
aiobotocore==2.5.4
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aioitertools==0.12.0
aiosignal==1.3.2
alabaster==0.7.16
annotated-types==0.7.0
anyio==4.1.0
app-model==0.3.1
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asciitree==0.3.3
asttokens==2.4.1
async-lru==2.0.4
async-timeout==5.0.1
attrs==23.1.0
babel==2.13.1
beautifulsoup4==4.12.2
bleach==6.1.0
blinker==1.9.0
botocore==1.31.17
bs4==0.0.1
build==1.2.2.post1
cachetools==5.3.2
cachey==0.2.1
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==2.1.1
click==8.1.7
cloudpickle==3.1.0
colorama==0.4.6
comm==0.2.0
configargparse==1.7
contourpy==1.2.0
cycler==0.12.1
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dask==2024.8.0
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
distributed==2024.8.0
docstring-parser==0.16
docutils==0.21.2
dominate==2.9.1
elementpath==4.7.0
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==2.0.1
fasteners==0.19
fastjsonschema==2.19.0
filelock==3.9.0
flask==3.0.3
flexcache==0.3
flexparser==0.4
fonttools==4.46.0
fqdn==1.5.1
freetype-py==2.5.1
frozenlist==1.5.0
fsspec==2023.6.0
google-auth==2.25.1
google-auth-oauthlib==1.1.0
grpcio==1.59.3
heapdict==1.0.1
hsluv==5.0.4
huggingface-hub==0.20.1
idna==3.4
imagecodecs==2023.9.18
imageio==2.33.1
imagesize==1.4.1
imglyb==2.1.0
importlib-metadata==7.0.0
importlib-resources==6.1.1
in-n-out==0.2.1
ipykernel==6.27.1
ipython==8.18.1
ipywidgets==8.1.5
isoduration==20.11.0
itk==5.4.0
itk-core==5.4.0
itk-elastix==0.20.0
itk-filtering==5.4.0
itk-io==5.4.0
itk-numerics==5.4.0
itk-registration==5.4.0
itk-segmentation==5.4.0
itsdangerous==2.2.0
jedi==0.19.1
jgo==1.0.6
jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jpype1==1.5.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter-client==8.6.0
jupyter-core==5.5.0
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter-server==2.12.0
jupyter-server-terminals==0.4.4
jupyterlab==4.0.9
jupyterlab-pygments==0.3.0
jupyterlab-server==2.25.2
jupyterlab-widgets==3.0.13
kiwisolver==1.4.5
labeling==0.1.14
lazy-loader==0.3
llvmlite==0.43.0
locket==1.0.0
lxml==4.9.4
magicgui==0.10.0
markdown==3.5.1
markdown-it-py==3.0.0
markupsafe==2.1.2
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
mpmath==1.2.1
msgpack==1.1.0
multidict==6.1.0
napari==0.5.5
napari-console==0.1.2
napari-plugin-engine==0.2.0
napari-svg==0.2.0
nbclient==0.9.0
nbconvert==7.12.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.0
notebook==7.0.6
notebook-shim==0.2.3
npe2==0.7.7
numba==0.60.0
numcodecs==0.12.1
numpy==1.24.1
numpydoc==1.8.0
oauthlib==3.2.2
ome-types==0.5.3
ome-zarr==0.10.2
open3d==0.18.0
opencv-python==4.7.0.72
openpyxl==3.1.2
overrides==7.4.0
packaging==23.2
pandas==2.1.3
pandocfilters==1.5.0
parso==0.8.3
partd==1.4.2
pillow==9.3.0
pint==0.24.4
platformdirs==4.1.0
plotly==5.24.1
pooch==1.8.2
probreg==0.3.8
prometheus-client==0.19.0
prompt-toolkit==3.0.41
propcache==0.2.1
protobuf==4.23.4
psutil==5.9.6
psygnal==0.11.1
ptflops==0.7.2.1
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pybind11==2.13.6
pyconify==0.1.6
pycparser==2.21
pydantic==2.10.4
pydantic-compat==0.1.2
pydantic-core==2.27.2
pygments==2.17.2
pyopengl==3.1.7
pyparsing==3.1.1
pyproject-hooks==1.2.0
python-dateutil==2.8.2
python-json-logger==2.0.7
pytorch-fid==0.3.0
pytorch-msssim==1.0.0
pytz==2023.3.post1
pywin32==306
pywinpty==2.0.12
pyyaml==6.0.1
pyzmq==25.1.2
qnorm==0.8.1
qtconsole==5.6.1
qtpy==2.4.2
referencing==0.31.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
resource-backed-dask-array==0.1.0
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
roifile==2024.9.15
rpds-py==0.13.2
rsa==4.9
s3fs==2023.6.0
safetensors==0.4.1
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.10.1
scyjava==1.10.2
seaborn==0.13.2
send2trash==1.8.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.4.1
sphinx==7.4.7
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
stack-data==0.6.3
superqt==0.7.0
sympy==1.11.1
tblib==3.0.0
tenacity==9.0.0
tensorboard==2.15.1
tensorboard-data-server==0.7.2
terminado==0.18.0
threadpoolctl==3.2.0
tifffile==2023.2.28
tinycss2==1.2.1
tokenizers==0.15.0
tomli==2.0.1
tomli-w==1.1.0
toolz==1.0.0
tornado==6.4
tqdm==4.66.1
traitlets==5.14.0
transforms3d==0.4.2
typer==0.15.1
types-python-dateutil==2.8.19.14
typing-extensions==4.12.2
tzdata==2023.3
uri-template==1.3.0
urllib3==1.26.13
vispy==0.14.3
wcwidth==0.2.12
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
werkzeug==3.0.1
widgetsnbextension==4.0.13
wrapt==1.17.0
xarray==2024.7.0
xmlschema==3.4.3
xsdata==24.3.1
yarl==1.18.3
zarr==2.15.0
zict==3.0.0
zipp==3.17.0